backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
mysql-connector-python==9.4.0
Flask-MySQLdb==2.0.0
mysqlclient==2.2.7
python-dotenv==1.1.1
pytest==7.4.0
flask-socketio==5.3.6
python-socketio==5.11.1